Fix compile error on arm.
On arm I get this error:
kword/part/KWView.cpp:1321: error: call of overloaded 'page(double)' is ambiguous
kword/part/KWPageManager.h:73: note: candidates are: KWPage KWPageManager::page(int) const
kword/part/KWPageManager.h:89: note:                 KWPage KWPageManager::page(qreal) const
so added explicit cast to qreal.

svn path=/trunk/koffice/; revision=1029944

+++ kword/part/KWView.cpp
@@ -1318,7 +1318,7 @@ void KWView::offsetInDocumentMoved(int yOffset)
             }
         }
     }
-    KWPage page = m_document->pageManager()->page(offset + height / 2.0);
+    KWPage page = m_document->pageManager()->page(qreal(offset + height / 2.0));
     if (page.isValid())
         setCurrentPage(page);
 }
